+++ CHANGES.rst
@@ -24,6 +24,7 @@ lqdn-gafam-campaign
 - Remove unknown or deprecated CSS- and SVG-properties
 - Add Makefile targets ``build-improved-resources`` and ``upload-mosaic``
 - Improve Python 3 compatibility for mosaic rendering
+- Fix broken "ý" glyphs with Czech language
 
 
 2021-12-31 0.11.0